Issue #753: In response to PR feedback: Updates return of type of list_registry function to List[str], Updates for loop in list_registry so function has single point of return Adds ak.EmptyRegistry and ak.EmptySymbolTable and updates references to the string literals '__EMPTY_REGISTRY__' and '__EMPTY_SYMBOLTABLE__' Moves test_string_registration and test_string_is_registered from string_test.py to registration_test.py
