+++ arkouda/client.py
@@ -8,8 +8,8 @@ from arkouda.logger import getArkoudaLogger
 from arkouda.message import RequestMessage, MessageFormat, ReplyMessage, \
      MessageType
 
-__all__ = ["AllSymbols", "RegisteredSymbols", "connect", "disconnect", "shutdown", "get_config",
-           "get_mem_used", "__version__", "ruok"]
+__all__ = ["AllSymbols", "RegisteredSymbols", "EmptySymbolTable", "EmptyRegistry", "connect", "disconnect", "shutdown",
+           "get_config", "get_mem_used", "__version__", "ruok"]
 
 # Try to read the version from the file located at ../VERSION
 VERSIONFILE = os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), 
@@ -40,7 +40,9 @@ pdarrayIterThresh  = pdarrayIterThreshDefVal
 maxTransferBytesDefVal = 2**30
 maxTransferBytes = maxTransferBytesDefVal
 AllSymbols = "__AllSymbols__"
+EmptySymbolTable = "__EMPTY_SYMBOLTABLE__"
 RegisteredSymbols = "__RegisteredSymbols__"
+EmptyRegistry = "__EMPTY_REGISTRY__"
 
 logger = getArkoudaLogger(name='Arkouda Client') 
 clientLogger = getArkoudaLogger(name='Arkouda User Logger', logFormat='%(message)s')
