+++ tests/string_test.py
@@ -472,50 +472,3 @@ class StringTest(ArkoudaTest):
         thickrange = thirds[0].stick(thirds[1], delimiter=', ').stick(thirds[2], delimiter=', ')
         flatrange = thickrange.flatten(', ')
         self.assertTrue((ak.cast(flatrange, 'int64') == ak.arange(99)).all())
-
-    def test_string_registration(self):
-        # Initial registration should set name
-        keep = ak.random_strings_uniform(1, 10, UNIQUE, characters='printable')
-        keep.register("keep_me")
-        self.assertTrue(keep.name is "keep_me")
-        self.assertTrue(keep.offsets.name == "keep_me_offsets")
-        self.assertTrue(keep.bytes.name == "keep_me_bytes")
-
-        # Register a second time to confirm name change
-        keep.register("kept")
-        self.assertTrue(keep.name is "kept")
-        self.assertTrue(keep.offsets.name == "kept_offsets")
-        self.assertTrue(keep.bytes.name == "kept_bytes")
-
-        # Add an item to discard
-        discard = ak.random_strings_uniform(1, 10, UNIQUE, characters='printable')
-
-        ak.clear()
-        self.assertTrue(keep.name == "kept")
-        self.assertTrue(keep.offsets.name == "kept_offsets")
-        self.assertTrue(keep.bytes.name == "kept_bytes")
-
-        with self.assertRaises(RuntimeError, msg="discard was not registered and should be discarded"):
-            str(discard)
-
-        # Unregister, should remain usable until we clear
-        keep.unregister()
-        str(keep) # Should not cause error
-        self.assertTrue(keep.offsets.name == "kept_offsets", msg="name should remain intact even after unregister")
-        ak.clear()
-        with self.assertRaises(RuntimeError, msg="keep was unregistered and should be cleared"):
-            str(keep) # should cause RuntimeError
-
-    def test_string_is_registered(self):
-        """
-        Tests the Strings.is_registered() function
-        """
-        keep = ak.random_strings_uniform(1, 10, UNIQUE, characters='printable')
-        self.assertFalse(keep.is_registered())
-
-        keep.register('keep_me')
-        self.assertTrue(keep.is_registered())
-
-        keep.unregister()
-        self.assertFalse(keep.is_registered())
-        ak.clear()
